Remove interface IMixer

+++ Sake/Analyzer.cs
@@ -60,7 +60,7 @@ namespace Sake
             return totalAnonsetWeighted / (decimal)sum;
         }
 
-        public static decimal BlockspaceEfficiency(IEnumerable<IEnumerable<decimal>> inputs, IEnumerable<IEnumerable<decimal>> outputs, long size)
+        public static decimal BlockspaceEfficiency(IEnumerable<IEnumerable<ulong>> inputs, IEnumerable<IEnumerable<ulong>> outputs, long size)
         {
             var avgAnon = AverageAnonsetGain(inputs, outputs);
             return avgAnon / (size / 1000m);
